requirements.txt
# Job Intelligence Agent
requests==2.32.3
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-telegram-bot==21.6
pypdf==4.3.1
tenacity==9.0.0
pydantic==2.9.2
ruff==0.6.9
